+++ lib/R5RS.scm
@@ -1298,15 +1298,7 @@
 
        Function return new Input Port with given filename. In Browser user need to
        provide global fs variable that is instance of FS interface."
-      (let ((fs (--> lips.env (get '**internal-env**) (get 'fs))))
-        (if (null? fs)
-            (throw (new Error "open-input-file: fs not defined"))
-            (begin
-              (if (not (procedure? readFile))
-                  (let ((_readFile (promisify fs.readFile)))
-                    (set! readFile (lambda (filename)
-                                     (--> (_readFile filename) (toString))))))
-              (new lips.InputFilePort (readFile filename) filename)))))))
+      (new lips.InputFilePort (%read-file false filename) filename))))
 
 ;; -----------------------------------------------------------------------------
 (define (close-input-port port)